# model/dto/product_dto/create_product_request.go
package product_dto

type CreateProductRequest struct {
	Name        string `json:"name"`
	Price       int    `json:"price"`
	Description string `json:"description"`
	Category    string `json:"category"`
	SellerId    int64  `json:"sellerId"`
}